--- a/src/Paprika/Data/SlottedArray.cs
+++ b/src/Paprika/Data/SlottedArray.cs
@@ -609,15 +609,17 @@ public readonly ref struct SlottedArray
                 return default;
             }
 
-            workingSet[0] = (byte)(hash >> HashByteShift);
-            NibblePath prefix;
-            if (count <= 4 && count > 2)
+            if (count <= 2 || count > 4)
             {
-                workingSet[1] = (byte)(hash & 0xFF);
+                workingSet[0] = (byte)(hash >> HashByteShift);
+            }
+            else
+            {
+                Unsafe.As<byte, ushort>(ref MemoryMarshal.GetReference(workingSet))
+                   = (ushort)((hash >> HashByteShift) | (hash << HashByteShift));
             }
 
-            prefix = NibblePath.FromKey(workingSet, 0, count > 4 ? KeySlice : count);
-
+            NibblePath prefix = NibblePath.FromKey(workingSet, 0, count > 4 ? KeySlice : count);
             if ((preamble & KeyPreambleOddBit) != 0)
             {
                 prefix.UnsafeMakeOdd(); // moving odd can make move beyond 0th