`UnPrepareKey` used in the enumeration over the map was slow when running `Importer`. This can also impact flushing data down while running the background `FlusherTask` of `Blockchain`. Should speed up data application as well.

### Benchmarks

#### Before

| Method       | Mean     | Error     | StdDev    | Code Size |
|------------- |---------:|----------:|----------:|----------:|
| EnumerateAll | 7.877 us | 0.1552 us | 0.1725 us |   2,487 B |

#### After

| Method       | Mean     | Error     | StdDev    | Code Size |
|------------- |---------:|----------:|----------:|----------:|
| EnumerateAll | 6.646 us | 0.0893 us | 0.0835 us |   1,756 B |