@Scooletz 2024-04-11
@benaadams I find no performance difference in the iterating scenario.